resolves #23: component registry 관리 포인트를 서비스 애플리케이션으로 이동
