+++ apps/code-generator/src/registry/meta/button.meta.tsx
@@ -2,7 +2,7 @@ import {
   COLOR_VARIANTS,
   SIZE_VARIANTS,
 } from "@packages/vanilla-extract-config";
-import { type ComponentMetaDefinition } from "../../types/meta";
+import { type ComponentMetaDefinition } from "../types";
 
 export const ButtonMeta: ComponentMetaDefinition = {
   component: "Button",
