+++ apps/code-generator/src/registry/meta/container.meta.tsx
@@ -1,5 +1,5 @@
 import { CONTAINER_WIDTH_SCALE } from "@packages/vanilla-extract-config";
-import { type ComponentMetaDefinition } from "../../types/meta";
+import { type ComponentMetaDefinition } from "../types";
 
 export const ContainerMeta: ComponentMetaDefinition = {
   component: "Container",
