+++ apps/code-generator/src/registry/meta/pagination.meta.tsx
@@ -2,7 +2,7 @@ import {
   SIZE_VARIANTS,
   THEME_COLOR_VARIANTS,
 } from "@packages/vanilla-extract-config";
-import { type ComponentMetaDefinition } from "../../types/meta";
+import { type ComponentMetaDefinition } from "../types";
 
 export const PaginationMeta: ComponentMetaDefinition = {
   component: "Pagination",
