+++ apps/code-generator/src/registry/meta/stack.meta.tsx
@@ -4,7 +4,7 @@ import {
   STACK_DIRECTION,
   STACK_JUSTIFICATION,
 } from "@packages/vanilla-extract-config";
-import { type ComponentMetaDefinition } from "../../types/meta";
+import { type ComponentMetaDefinition } from "../types";
 
 const Item = ({ index }: { index: number }) => (
   <div
