+++ packages/ui/src/components/ui/badge.tsx
@@ -1,6 +1,13 @@
 import type { Ref, HTMLAttributes } from "react";
 import { cn } from "../../utils/cn";
-import { type BadgeVariants, badge } from "./badge.css";
+import {
+  type BadgeVariants,
+  badge,
+  BADGE_COLOR_VARIANTS,
+  BADGE_SHAPES,
+} from "./badge.css";
+
+export { BADGE_COLOR_VARIANTS, BADGE_SHAPES };
 
 export type BadgeProps = HTMLAttributes<HTMLSpanElement> & {
   ref?: Ref<HTMLSpanElement>;
