+++ packages/ui/src/index.ts
@@ -1,30 +1,41 @@
 // Theme css
 import "./styles/theme.css";
 
-// Component Types
-export type {
-  ComponentMetaDefinition,
-  PropsMeta,
-  ControlType,
-} from "./types/meta";
-
-// Component Registry
+// Component export
+export { Container } from "./components/layout/container";
+export { Section } from "./components/layout/section";
+export { Stack } from "./components/layout/stack";
+export { Grid } from "./components/layout/grid";
 export {
-  TableComponentRegistry,
-  ComponentRegistry,
-  ComponentsByCategory,
-  getComponent,
-  getComponentMeta,
-  getAllComponents,
-} from "./registry";
-export type {
-  ComponentName,
-  ComponentMeta,
-  ComponentRegistryItem,
-  ComponentType,
-  ComponentCategoryName,
-  ComponentItem,
-} from "./registry";
+  Typography,
+  TYPOGRAPHY_ELEMENT,
+} from "./components/display/typography";
+export { Pagination } from "./components/display/pagination";
+export {
+  Table,
+  Tbody,
+  Thead,
+  TableCaption,
+  Tr,
+  Td,
+  Th,
+} from "./components/display/table";
+export { Button } from "./components/ui/button";
+export { Divider } from "./components/ui/divider";
+export { Card } from "./components/ui/card";
+export {
+  Badge,
+  BADGE_COLOR_VARIANTS,
+  BADGE_SHAPES,
+} from "./components/ui/badge";
+export { Avatar, AVATAR_STATUS } from "./components/ui/avatar";
+export { Input } from "./components/inputs/input";
+export { Checkbox } from "./components/inputs/checkbox";
+export { Radio } from "./components/inputs/radio";
+export { Textarea } from "./components/inputs/textarea";
+export { Select } from "./components/inputs/select";
+export { FormField } from "./components/inputs/form-field";
+export { FormLabel } from "./components/inputs/form-label";
 
 // class generating utils
 export { cn } from "./utils/cn";
